README.md
This is an app that is designed to help students learn better Mathematics. 
If the answer is right, it adds 2 points to your overall score. 
Your total determines the message displayed. 
If all answers are correct the student is congratulated when the answer is wrong the student is encouraged.

Find the App here 
https://drive.google.com/open?id=1yu9qRQoIFvCWcojEmZV9Um4B9G7eyBDh